Rename directory startScan to scaPage
