+++ Android/app/src/main/java/fr/uge/structsure/database/AppDatabase.kt
@@ -8,12 +8,12 @@ import androidx.room.migration.Migration
 import androidx.sqlite.db.SupportSQLiteDatabase
 import fr.uge.structsure.connexionPage.data.AccountDao
 import fr.uge.structsure.connexionPage.data.AccountEntity
-import fr.uge.structsure.startScan.data.ResultSensors
-import fr.uge.structsure.startScan.data.ScanEntity
-import fr.uge.structsure.startScan.data.StructureEntity
-import fr.uge.structsure.startScan.data.dao.ResultDao
-import fr.uge.structsure.startScan.data.dao.ScanDao
-import fr.uge.structsure.startScan.data.dao.StructurePlanDao
+import fr.uge.structsure.scanPage.data.ResultSensors
+import fr.uge.structsure.scanPage.data.ScanEntity
+import fr.uge.structsure.scanPage.data.StructureEntity
+import fr.uge.structsure.scanPage.data.dao.ResultDao
+import fr.uge.structsure.scanPage.data.dao.ScanDao
+import fr.uge.structsure.scanPage.data.dao.StructurePlanDao
 import fr.uge.structsure.structuresPage.data.PlanDB
 import fr.uge.structsure.structuresPage.data.PlanDao
 import fr.uge.structsure.structuresPage.data.SensorDB
