+++ Android/app/src/main/java/fr/uge/structsure/scanPage/domain/TimedBuffer.kt
@@ -1,4 +1,4 @@
-package fr.uge.structsure.startScan.domain
+package fr.uge.structsure.scanPage.domain
 
 import android.os.SystemClock
 import android.util.Log
