+++ Android/app/src/main/java/fr/uge/structsure/scanPage/presentation/DisplayPlanImages.kt
@@ -1,11 +1,5 @@
-package fr.uge.structsure.startScan.presentation
+package fr.uge.structsure.scanPage.presentation
 
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.layout.*
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.res.painterResource
 /*
 @Composable
 fun DisplayPlanImages(plans: List<PlanEntity>) {
